Support deploying sylva-core
Support optionally deploy sylva-core components. To enable
automatically deploying a workload cluster from Flux using
a private repo with a custom trusted HTTPs certificate, we
have to explicitly create the Flux GitRepository with the
secretRef containing the trusted CA certificate. This is
due to a restriction in Sylva which disallows specifying
the "secretRef" in the GitRepository spec.