Configure additional storage classes
This commit installs additional provisioner
with multiple exports from a single NFS server.
The /nfs/share path which existed earlier used
to configure the "dynamic" storage class for
ironic-shared-volume and the /var/lib/mysql
path is used to configure the "mysql" storage class
which is used for the mariadb datastore.
Permissions on these paths are configured differently.